Move C related files to C folder
