+++ ffcx/codegeneration/C/finite_element.py
@@ -11,7 +11,7 @@
 import logging
 
 import ffcx.codegeneration.basix_custom_element_template as ufcx_basix_custom_finite_element
-import ffcx.codegeneration.finite_element_template as ufcx_finite_element
+import ffcx.codegeneration.C.finite_element_template as ufcx_finite_element
 import ufl
 
 logger = logging.getLogger("ffcx")
